-- Fixed the message about deleting non-exisiting validation.log    files (was generated in the top level validate.sh).
-- Created a SolidTetMesh in tetgen_mesh.template.* Renamed
   the many copies thereof in various driver codes to avoid
   clashes.

-- Created a new src directory

          mesh_smoothing

   which contains a single header file that contains the
   functors that can be used to smooth meshes whose
   boundary nodes have been moved. (Smoothing can be done
   by solving a Poisson equation, linear and nonlinear 
   elasticity).

-- Moved all the vmtk mesh generation and transfer to quadratic
   boundary codes to  

         demo_drivers/meshing/quadratic_vmtk_tetgen


   which contains a script that illustrates how to create
   quadratic mesh information from a vmtk-generated *.vtp
   file. The driver code snap_mesh.cc then shows how to
   smooth the fluid and solid meshes in which the boundary
   nodes are snapped onto the quadratic boundaries.


-- Cleaned up the  

        demo_drivers/interaction/unsteady_vmtk_fsi

   directory. It now only contains the unsteady vmtk-based
   FSI code, using the quadratic-ified geometry.




git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@118 595e1d82-adf1-4282-bac7-b6073fa2ca8e

+++ validate.sh
@@ -25,7 +25,7 @@ OptionRead()
 echo " "
 echo "================================================================="
 echo " "
-echo "I'm about to run self-tests." 
+echo "I'm about to run the self-tests." 
 echo " "
 echo "Note: The self-test suite involves a large number of separate  "
 echo "      test scripts in different directories. We will process all of "
@@ -44,14 +44,15 @@ echo "     it was passed succesfully! "
 echo " "
 echo "================================================================="
 echo " "
-echo " I will wipe the existing validation.log file."
-echo " "
-OptionPrompt " Is this OK? [y/n]"
-reply=`OptionRead`
-if test "$reply" != "y" -a "$reply" != "Y" ; then 
-   OptionPrompt "Terminating..."
-   exit
+if [ -e validation.log ]; then
+     echo " I will wipe the existing validation.log file."
+     echo " "
+     OptionPrompt " Is this OK? [y/n]"
+     reply=`OptionRead`
+     if test "$reply" != "y" -a "$reply" != "Y" ; then 
+         OptionPrompt "Terminating..."
+         exit
+     fi
+     rm -f validation.log
 fi
-rm -f validation.log
-
 exit 0
